get rid of sub conftest and folder
